pyproject.toml
[project]
name = "huproof"
version = "0.1.0"
description = "Human-in-the-Loop ZK Auth PoC (privacy-preserving, unlinkable liveness)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Your Name" }]
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.2.1",
    "sqlmodel>=0.0.22",
    "sqlalchemy>=2.0.25",
    "structlog>=24.1.0",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.9",
    "typing-extensions>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "pytest>=8.2.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"




